requirements.txt
SQLAlchemy==2.0.35
alembic==1.13.3
asyncpg==0.29.0
fastapi==0.115.2
uvicorn==0.31.1
pydantic==2.9.2
pydantic_core==2.23.4
pydantic[email]
httpx==0.27.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.0
python-multipart==0.0.12
python-dotenv==1.0.1
pytest-asyncio==0.24.0
pytest==8.3.3